requirements.txt
behave==1.2.6
google-auth==2.36.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.9.0
requests==2.32.3
urllib3==2.2.3